+++ crates/token/Cargo.toml
@@ -30,8 +30,11 @@ namada_events = { path = "../events", default-features = false }
 namada_macros = { path = "../macros" }
 namada_migrations = { path = "../migrations", optional = true }
 namada_shielded_token = { path = "../shielded_token" }
+namada_storage = { path = "../storage" }
 namada_systems = { path = "../systems" }
 namada_trans_token = { path = "../trans_token" }
+namada_tx = { path = "../tx" }
+namada_tx_env = { path = "../tx_env" }
 
 arbitrary = { workspace = true, optional = true }
 borsh.workspace = true
